+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- Match: a fuzzy matcher module which is used to fuzzy match a query within a list of provided user items
 - Select: user interface providing an interactive set of components to input, display and preview items within a user defined list
-- Match: a matcher module which is used to fuzzy match a query within a list of provided user items
-- Stream: an async stream component which can start system utilities or stream from in-memory sources in a non-blocking interactive way
+- Stream: async stream component which can start system utilities or stream from in-memory sources in a non-blocking interactive way
+- Picker: a component which combines the core Select, Match and Stream and provides a rich extensible interface to create dynamic user pickers.